+++ base_layer/wallet/src/util/mod.rs
@@ -22,5 +22,6 @@
 
 pub mod diesel_ext;
 pub mod encryption;
+pub mod one_sided;
 pub mod wallet_identity;
 pub mod watch;
